Allow specifying command to require user to be bot admin
Bot admins are set via `.env` file and is string of discord user IDs separated by comma.

Question I have for this approach now is if the admins should be defined with discord user tags, e.g. user#1234, instead of 18 number long user ID. Would be easier to use the tag but is it less secure? Tag can be changed but I assume id cannot so using id could be more secure.