Remove content parameter from commands run function
For now returning `removeTrigger` function to allow users to remove the trigger if they want. It was unclear that the `content` would have the message without the trigger word. If you had e.g. trigger with `hello` and the message was "This is something before hello. Extra words" and the value of `content` would have been "This is something before. Extra words".

Returning the function itself to remove the trigger feels more clear.